refactor(core/dialog): migrate recipe to config slot recipe
Move Dialog from sva() to defineSlotRecipe registered in panda.config,
matching the Card pattern for consistency across compound components.
