+++ src/recipes/index.ts
@@ -1,9 +1,11 @@
 import { card } from './card';
+import { dialog } from './dialog';
 import { icon } from './icon';
 import { menu } from './menu';
 
 export const slotRecipes = {
   card,
+  dialog,
   menu,
 };
